% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vals.R
\name{extract_vals}
\alias{extract_vals}
\title{Extract True and Estimated Values
Function that extracts time series of quantities (truth and estimates) from \code{\link{read_rep}}}
\usage{
extract_vals(replist, quantname, nsim = Nsim)
}
\arguments{
\item{quantname}{El nombre de la cantidad para buscar en el archivo report}

\item{nsim}{El numero de replicas}

\item{truerep}{Una lista de \code{\link{read_rep}} que es del modelo baso}

\item{allreps}{Una lista de listas con valores estimados de los modelos estimaciones}
}
\value{
Named list of "tru" and "est" values
}
\description{
Extract True and Estimated Values
Function that extracts time series of quantities (truth and estimates) from \code{\link{read_rep}}
}
\examples{
\dontrun{
reps <- get_reps(OMfolder,simrepfile,truerepfile)
extract_vals(reps,'Biomasatotal')
}
}
