% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reps.R
\name{get_reps}
\alias{get_reps}
\title{Read in Report Files
Tiene que estar en la carpeta de "Dropbox/Simulador\ modelo\ Inpesca"}
\usage{
get_reps(OMfolder, simrepfile, truerepfile)
}
\arguments{
\item{OMfolder}{Folder containing runs}

\item{simrepfile}{Base name of each simulation rep file (.}

\item{truerepfile}{final year to plot from. Default value is 2018.}
}
\value{
Rdata file with \code{allreps} and \code{truerep}
}
\description{
Read in Report Files
Tiene que estar en la carpeta de "Dropbox/Simulador\ modelo\ Inpesca"
}
\examples{
\dontrun{
	OMfolder <- "Evalua\\ Estimador_INPESCA_SARD"
simrepfile <- "sar_sim" # "mae_sim"
truerepfile <- "sar_base.rep"
 reps <- get_reps(OMfolder,simrepfile,truerepfile)
}
}
