% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_REs.R
\name{get_REs}
\alias{get_REs}
\title{Get Relative Errors
Function that plots time series of relative errors based on output from \code{\link{extract_vals}}}
\usage{
get_REs(vals, endyr = 2018, Nyears = 3)
}
\arguments{
\item{vals}{list output from \code{\link{extract_vals}}}

\item{endyr}{final year to plot from. Default value is 2019.}

\item{Nyears}{number of final years to calculate for}
}
\value{
Data frame of eight values
}
\description{
Get Relative Errors
Function that plots time series of relative errors based on output from \code{\link{extract_vals}}
}
\examples{
\dontrun{
plot_RE(extract_vals(reps,'Biomasatotal'))
}
}
