% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RE.R
\name{plot_RE}
\alias{plot_RE}
\title{Plot Relative Errors
Function that plots time series of relative errors based on output from \code{\link{extract_vals}}}
\usage{
plot_RE(vals, qname, endyr = 2018)
}
\arguments{
\item{vals}{list output from \code{\link{extract_vals}}}

\item{endyr}{final year to plot from. Default value is 2018.}
}
\value{
Box and whiskers plot of relative errors
}
\description{
Plot Relative Errors
Function that plots time series of relative errors based on output from \code{\link{extract_vals}}
}
\examples{
\dontrun{
plot_RE(extract_vals(reps,'Biomasatotal'))
}
}
