% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_REs.R
\name{plot_REs}
\alias{plot_REs}
\title{Plot Relative Errors
Function that plots time series of relative errors based on output from \code{\link{extract_vals}}}
\usage{
plot_REs(vals, qname, endyr = 2019)
}
\arguments{
\item{vals}{list output from \code{\link{extract_vals}}}

\item{qname}{name of quantity being plotted}

\item{endyr}{final year to plot from. Default value is 2019.}
}
\value{
Box and whiskers plot of relative errors
}
\description{
Plot Relative Errors
Function that plots time series of relative errors based on output from \code{\link{extract_vals}}
}
\examples{
\dontrun{
plot_RE(extract_vals(reps,'Biomasatotal'))
}
}
