% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_extract_vals.R
\name{SS_extract_vals}
\alias{SS_extract_vals}
\title{Extract values from SS report files
Function that extracts quantities from SSsummarize output}
\usage{
SS_extract_vals(repfiles, quantname)
}
\arguments{
\item{repfiles}{Una lista de \code{\link{SSsummarize}}}

\item{quantname}{El nombre de la cantidad para buscar en el archivo report}
}
\value{
List of tru and est values
}
\description{
Extract values from SS report files
Function that extracts quantities from SSsummarize output
}
\examples{
\dontrun{
plot_RE(extract_vals(reps,'Biomasatotal'))
}
}
