% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_get_reps.R
\name{SS_get_reps}
\alias{SS_get_reps}
\title{Efficiently Read in Report Files from SS
Tiene que estar en la carpeta de "Dropbox/Simulador\ modelo\ Inpesca"
Por ahora solo esta}
\usage{
SS_get_reps(OMfolder)
}
\arguments{
\item{OMfolder}{Folder containing runs}

\item{simrepfile}{Base name of each simulation rep file (.}

\item{truerepfile}{final year to plot from. Default value is 2018.}
}
\value{
Rdata file with \code{allreps} and \code{truerep}
}
\description{
Efficiently Read in Report Files from SS
Tiene que estar en la carpeta de "Dropbox/Simulador\ modelo\ Inpesca"
Por ahora solo esta
}
\examples{
\dontrun{
	OMfolder <- "Evalua\\ Estimador_INPESCA_SARD"
simrepfile <- "sar_sim" # "mae_sim"
truerepfile <- "sar_base.rep"
 reps <- get_reps(OMfolder,simrepfile,truerepfile)
}
}
